+++ pyuploadcare/secure_url.py
@@ -12,6 +12,11 @@ class BaseSecureUrlBuilder(ABC):
     def build(self, uuid: str, wildcard: bool = False) -> str:
         raise NotImplementedError
 
+    def get_token(self, uuid: str, wildcard: bool = False) -> str:
+        raise NotImplementedError(
+            f"{self.__class__} doesn't provide get_token()"
+        )
+
 
 class BaseAkamaiSecureUrlBuilder(BaseSecureUrlBuilder):
     """Akamai secure url builder.
@@ -20,6 +25,7 @@ class BaseAkamaiSecureUrlBuilder(BaseSecureUrlBuilder):
     for more details.
     """
 
+    template = "{base}?token={token}"
     field_delimeter = "~"
 
     def __init__(
@@ -34,15 +40,31 @@ class BaseAkamaiSecureUrlBuilder(BaseSecureUrlBuilder):
         self.window = window
         self.hash_algo = hash_algo
 
+    def build(self, uuid: str, wildcard: bool = False) -> str:
+        uuid_or_url = self._format_uuid_or_url(uuid)
+        token = self.get_token(uuid_or_url, wildcard=wildcard)
+        secure_url = self._build_url(uuid_or_url, token)
+        return secure_url
+
+    def get_token(self, uuid: str, wildcard: bool = False) -> str:
+        uuid_or_url = self._format_uuid_or_url(uuid)
+        expire = self._build_expire_time()
+        acl = self._format_acl(uuid_or_url, wildcard=wildcard)
+        signature = self._build_signature(uuid_or_url, expire, acl)
+        token = self._build_token(expire, acl, signature)
+        return token
+
     def _build_expire_time(self) -> int:
         return int(time.time()) + self.window
 
     def _build_signature(
-        self, expire: int, acl: Optional[str] = None, url: Optional[str] = None
+        self, uuid_or_url: str, expire: int, acl: Optional[str]
     ) -> str:
-        assert bool(acl) != bool(url)
 
-        hash_source = [f"exp={expire}", f"acl={acl}" if acl else f"url={url}"]
+        hash_source = [
+            f"exp={expire}",
+            f"acl={acl}" if acl else f"url={uuid_or_url}",
+        ]
 
         signature = hmac.new(
             binascii.a2b_hex(self.secret_key.encode()),
@@ -52,101 +74,72 @@ class BaseAkamaiSecureUrlBuilder(BaseSecureUrlBuilder):
 
         return signature
 
+    def _build_token(self, expire: int, acl: Optional[str], signature: str):
 
-class AkamaiSecureUrlBuilderWithAclToken(BaseAkamaiSecureUrlBuilder):
-    template = "https://{cdn}/{uuid}/?token={token}"
+        token_parts = [
+            f"exp={expire}",
+            f"acl={acl}" if acl else None,
+            f"hmac={signature}",
+        ]
 
-    def build(self, uuid: str, wildcard: bool = False) -> str:
-        uuid = uuid.lstrip("/").rstrip("/")
+        return self.field_delimeter.join(
+            part for part in token_parts if part is not None
+        )
 
-        expire = self._build_expire_time()
-
-        acl = self._format_acl(uuid, wildcard=wildcard)
-
-        signature = self._build_signature(expire, acl=acl)
-
-        secure_url = self._build_url(uuid, expire, acl, signature)
-        return secure_url
+    @abstractmethod
+    def _build_base_url(self, uuid_or_url: str):
+        raise NotImplementedError
 
     def _build_url(
         self,
-        uuid: str,
-        expire: int,
-        acl: str,
-        signature: str,
+        uuid_or_url: str,
+        token: str,
     ) -> str:
-        req_parameters = [
-            f"exp={expire}",
-            f"acl={acl}",
-            f"hmac={signature}",
-        ]
-
-        token = self.field_delimeter.join(req_parameters)
-
+        base_url = self._build_base_url(uuid_or_url)
         return self.template.format(
-            cdn=self.cdn_url,
-            uuid=uuid,
+            base=base_url,
             token=token,
         )
 
-    def _build_token(self, expire: int, acl: Optional[str], signature: str):
-        token_parts = [
-            f"exp={expire}",
-            f"acl={acl}",
-            f"hmac={signature}",
-        ]
-        return self.field_delimeter.join(token_parts)
+    @abstractmethod
+    def _format_acl(self, uuid_or_url: str, wildcard: bool) -> Optional[str]:
+        raise NotImplementedError
 
-    def _format_acl(self, uuid: str, wildcard: bool) -> str:
+    @abstractmethod
+    def _format_uuid_or_url(self, uuid_or_url: str) -> str:
+        raise NotImplementedError
+
+
+class AkamaiSecureUrlBuilderWithAclToken(BaseAkamaiSecureUrlBuilder):
+    base_template = "https://{cdn}/{uuid}/"
+
+    def _build_base_url(self, uuid_or_url: str):
+        return self.base_template.format(cdn=self.cdn_url, uuid=uuid_or_url)
+
+    def _format_acl(self, uuid_or_url: str, wildcard: bool) -> str:
         if wildcard:
-            return f"/{uuid}/*"
-        return f"/{uuid}/"
+            return f"/{uuid_or_url}/*"
+        return f"/{uuid_or_url}/"
+
+    def _format_uuid_or_url(self, uuid_or_url: str) -> str:
+        return uuid_or_url.lstrip("/").rstrip("/")
 
 
 class AkamaiSecureUrlBuilderWithUrlToken(BaseAkamaiSecureUrlBuilder):
-    template = "{url}?token={token}"
+    def _build_base_url(self, uuid_or_url: str):
+        return uuid_or_url
 
-    def build(self, uuid: str, wildcard: bool = False) -> str:
+    def _format_acl(self, uuid_or_url: str, wildcard: bool) -> None:
         if wildcard:
             raise ValueError(
                 "Wildcards are not supported in AkamaiSecureUrlBuilderWithUrlToken."
             )
+        return None
 
-        url = uuid
-
-        expire = self._build_expire_time()
-
-        signature = self._build_signature(expire, url=url)
-
-        secure_url = self._build_url(url, expire, signature)
-
-        return secure_url
-
-    def _build_url(
-        self,
-        url: str,
-        expire: int,
-        signature: str,
-    ) -> str:
-        req_parameters = [
-            f"exp={expire}",
-            f"hmac={signature}",
-        ]
-
-        token = self.field_delimeter.join(req_parameters)
-
-        return self.template.format(
-            url=url,
-            token=token,
-        )
-
-    def _build_token(self, expire: int, url: str, signature: str):
-        token_parts = [
-            f"exp={expire}",
-            f"url={url}",
-            f"hmac={signature}",
-        ]
-        return self.field_delimeter.join(token_parts)
+    def _format_uuid_or_url(self, uuid_or_url: str) -> str:
+        if "://" not in uuid_or_url:
+            raise ValueError(f"{uuid_or_url} doesn't look like a URL")
+        return uuid_or_url
 
 
 class AkamaiSecureUrlBuilder(AkamaiSecureUrlBuilderWithAclToken):