+++ tests/functional/test_secure_url.py
@@ -13,6 +13,19 @@ known_secret = (
 
 
 @pytest.mark.freeze_time("2021-10-12")
+def test_get_acl_token():
+    secure_url_bulder = AkamaiSecureUrlBuilderWithAclToken(
+        "cdn.yourdomain.com", known_secret
+    )
+    token = secure_url_bulder.get_token("52da3bfc-7cd8-4861-8b05-126fef7a6994")
+    assert token == (
+        "exp=1633997100~"
+        "acl=/52da3bfc-7cd8-4861-8b05-126fef7a6994/~"
+        "hmac=81852547d9dbd9eefd24bee2cada6eab02244b9013533bc8511511923098df72"
+    )
+
+
+@pytest.mark.freeze_time("2021-10-12")
 def test_generate_secure_url_acl_token():
     secure_url_bulder = AkamaiSecureUrlBuilderWithAclToken(
         "cdn.yourdomain.com", known_secret
@@ -80,6 +93,41 @@ def test_client_generate_secure_url_acl_token():
         "exp=1633997100~"
         "acl=/52da3bfc-7cd8-4861-8b05-126fef7a6994/~"
         "hmac=81852547d9dbd9eefd24bee2cada6eab02244b9013533bc8511511923098df72"
+    )
+
+
+@pytest.mark.freeze_time("2021-10-12")
+def test_client_generate_secure_url_token_acl_token():
+    secure_url_bulder = AkamaiSecureUrlBuilderWithAclToken(
+        "cdn.yourdomain.com", known_secret
+    )
+
+    uploadcare = Uploadcare(
+        public_key="public",
+        secret_key="secret",
+        secure_url_builder=secure_url_bulder,
+    )
+    secure_url = uploadcare.generate_secure_url_token(
+        "52da3bfc-7cd8-4861-8b05-126fef7a6994"
+    )
+    assert secure_url == (
+        "exp=1633997100~"
+        "acl=/52da3bfc-7cd8-4861-8b05-126fef7a6994/~"
+        "hmac=81852547d9dbd9eefd24bee2cada6eab02244b9013533bc8511511923098df72"
+    )
+
+
+@pytest.mark.freeze_time("2021-10-12")
+def test_get_url_token():
+    secure_url_bulder = AkamaiSecureUrlBuilderWithUrlToken(
+        "cdn.yourdomain.com", known_secret
+    )
+    token = secure_url_bulder.get_token(
+        "https://cdn.yourdomain.com/52da3bfc-7cd8-4861-8b05-126fef7a6994/"
+    )
+    assert token == (
+        "exp=1633997100~"
+        "hmac=32b696b855ddc911b366f11dcecb75789adf6211a72c1dbdf234b83f22aaa368"
     )